requirements.txt
torch
torchvision
timm
pandas
pillow
scikit-learn
matplotlib
numpy